gkeys/lib.py: Add saving the seed to the installed keydir
This will make knowing what the seed was that was used to install the key.
It will make searching the installed db easier, by loading the saved seed.
Searches on the keys will then be able to use these saved seeds to search for something 
other that a nick.
It will make comparing the remote seed file with the installed seed easier by using 
the GKEY's __cmp__().